Modify store & name filename generation

+++ src/vxfs/libs/snowflake.go
@@ -47,7 +47,7 @@ func NewSnowFlake(machineId int64) (*SnowFlake, error) {
 	return i, nil
 }
 
-func (i *SnowFlake) UnsafeNextId() (uint64, error) {
+func (i *SnowFlake) UnsafeNextId() (int64, error) {
 	timestamp := twTimestamp()
 	if timestamp < i.lastTimestamp {
 		return 0, errors.New(fmt.Sprintf("Clock moved backwards. Refusing %d milliseconds", i.lastTimestamp-timestamp))
@@ -61,16 +61,16 @@ func (i *SnowFlake) UnsafeNextId() (uint64, error) {
 		i.sequence = 0
 	}
 	i.lastTimestamp = timestamp
-	return uint64((timestamp << 22) | (i.machineId << 12) | i.sequence), nil
+	return (timestamp << 22) | (i.machineId << 12) | i.sequence, nil
 }
 
-func (i *SnowFlake) NextId() (uint64, error) {
+func (i *SnowFlake) NextId() (int64, error) {
 	i.safeLock.Lock()
 	defer i.safeLock.Unlock()
 	return i.UnsafeNextId()
 }
 
-func (i *SnowFlake) NextIds(count int) ([]uint64, error) {
+func (i *SnowFlake) NextIds(count int) ([]int64, error) {
 	if count < 1 || count > maxSequenceMask {
 		return nil, errors.New(fmt.Sprintf("NextIds count error, limit to 1 ~ %d", maxSequenceMask))
 	}
@@ -80,7 +80,7 @@ func (i *SnowFlake) NextIds(count int) ([]uint64, error) {
 
 	var (
 		err error
-		ids = make([]uint64, count)
+		ids = make([]int64, count)
 	)
 	for n := 0; n < count; n++ {
 		if ids[n], err = i.UnsafeNextId(); err != nil {
